requirements.txt
python-dotenv
pillow
anthropic
requests
aiohttp
pydub
